backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.4
redis==6.4.0
celery==5.5.3
slowapi==0.1.9
loguru==0.7.3
prometheus-fastapi-instrumentator==7.1.0
python-magic==0.4.27
opencv-python-headless==4.9.0.80
pillow==11.3.0
scikit-image==0.25.2
pydicom==3.0.1
numpy==1.26.4
tensorflow==2.16.1
tf-keras-vis==0.8.7
lime==0.2.0.1
shap==0.48.0
reportlab==4.4.3
minio==7.2.16
boto3==1.34.67
botocore==1.33.11
httpx==0.28.1
pytest==8.4.2
pytest-asyncio==1.1.0
ruff==0.12.12
mypy==1.17.1
email-validator==2.3.0
